prass-app: move conduit folder from components to pages
add new route for conduit in main.js
