--- praas-app/src/app/components/conduit/index.js
@@ -1,4 +1,0 @@
-import ConduitList from './list';
-import { CreateConduitForm, EditConduitForm } from './form';
-
-export { ConduitList, CreateConduitForm, EditConduitForm };
