+++ praas-app/src/app/main.js
@@ -18,6 +18,7 @@ import configureStore from 'store';
 import Home from './pages/home';
 import Login from './pages/login';
 import Signup from './pages/signup';
+import Conduit from './pages/conduit';
 
 // TODO:
 // - clear alert on location change
@@ -28,6 +29,7 @@ export default function App(props) {
       <Route path="/" element={<Home />} />
       <Route path="login" element={<Login />} />
       <Route path="signup" element={<Signup />} />
+      <Route path="conduit" element={<Conduit />} />
     </Routes>
   );
 }
